+++ scripts/camotics-init-dev-img.sh
@@ -8,4 +8,4 @@ apt-get update
 #apt-get dist-upgrade -y
 
 # Install packages
-apt-get install -y scons build-essential libssl-dev python3-dev
+apt-get install -y scons build-essential libssl-dev python3-dev nodejs-dev