+++ src/js/axis-vars.js
@@ -115,7 +115,7 @@ module.exports = {
       case 'UNDER':
         title = 'Current program would move ' +
           this._length_str(min - pathMin - off) + ' below axis bounds.  ' +
-          'Offsets must be adjusted of soft limits set correctly.';
+          'Offsets must be adjusted or soft limits set correctly.';
         break;
 
       case 'NO FIT':